% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas.R
\name{apply_predicate_to_results}
\alias{apply_predicate_to_results}
\title{Utility function for row-wise filtering of results data frame.}
\usage{
apply_predicate_to_results(results, predicate)
}
\description{
Utility function for row-wise filtering of results data frame.
}
