% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas.R
\docType{class}
\name{BinaryConfiguration-class}
\alias{BinaryConfiguration-class}
\alias{BinaryConfiguration}
\title{A configuration class that is specific to binary traits pheWAS.}
\description{
A configuration class that is specific to binary traits pheWAS.
}
