% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{code_in_range}
\alias{code_in_range}
\title{Tests if a given code is in an inclusive range of codes.}
\usage{
code_in_range(code, left, right)
}
\arguments{
\item{code}{The code to test.}

\item{left}{The left bound of the inclusive range.}

\item{right}{The right bound of the inclusive range.}
}
\value{
T if the code is in the range, F otherwise
}
\description{
Tests if a given code is in an inclusive range of codes.
}
