% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{codes_in_range}
\alias{codes_in_range}
\title{Check if an array of ICD10 codes fall within a range.}
\usage{
codes_in_range(codes, left, right)
}
\description{
Check if an array of ICD10 codes fall within a range.
}
