% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{compare_codes}
\alias{compare_codes}
\title{Comparison operator for ICD10 codes.}
\usage{
compare_codes(a, b)
}
\arguments{
\item{a}{The first ICD10 code.}

\item{b}{The second ICD10 code.}
}
\value{
-1 if a < b, 0 if a == b and 1 if a > b
}
\description{
Utility function for ICD10 ordering so that it is possible to quickly
check if a code is within a range of codes.
This function is aware of the code structure (e.g. so that say B02 would be
greater than A05).
If a code is a subcode of another one (e.g. I20.0 and I20), the return value
will be 0.
}
