% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R, R/phewas.R
\docType{class}
\name{Configuration-class}
\alias{Configuration-class}
\alias{Configuration}
\title{Helper function to create Configuration objects.}
\description{
Helper function to create Configuration objects.

A configuration class to parametrize PheWAS analyses.
}
