% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R, R/phewas.R
\docType{class}
\name{ContinuousConfiguration-class}
\alias{ContinuousConfiguration-class}
\alias{ContinuousConfiguration}
\title{A configuration class that is specific to continuous traits pheWAS.}
\description{
A configuration class that is specific to continuous traits pheWAS.

A configuration class that is specific to continuous traits pheWAS.
}
