% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{fastglm_anova}
\alias{fastglm_anova}
\title{Analysis of deviance test for a fit based on fastglm.}
\usage{
fastglm_anova(base, augmented)
}
\description{
Analysis of deviance test for a fit based on fastglm.
}
