% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas.R
\name{format_logistic_fit}
\alias{format_logistic_fit}
\title{Format a GLM fit object.}
\usage{
format_logistic_fit(fit, data, fastglm_format)
}
\description{
This function calculates some extra statistics and formats the fit object
as a dataframe.
}
