% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{generator_cv_endpoints}
\alias{generator_cv_endpoints}
\title{Generator for the manually defined cv_endpoints.}
\usage{
generator_cv_endpoints(configuration, raw_data, callback, cl, limit = NULL)
}
\description{
Generator for the manually defined cv_endpoints.
}
