% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{generator_icd10_blocks}
\alias{generator_icd10_blocks}
\title{Generate phenotypes based on ICD10 blocks.}
\usage{
generator_icd10_blocks(configuration, raw_data, callback, cl,
  limit = NULL)
}
\description{
Generate phenotypes based on ICD10 blocks.
}
