% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{generator_icd10_raw}
\alias{generator_icd10_raw}
\title{Generate phenotypes based on raw ICD10 codes.}
\usage{
generator_icd10_raw(configuration, raw_data, callback, cl, limit = NULL)
}
\description{
Generate phenotypes based on raw ICD10 codes.
}
