% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{generator_icd10_three_chars}
\alias{generator_icd10_three_chars}
\title{Generate phenotypes based on 3 character ICD10 codes.}
\usage{
generator_icd10_three_chars(
  configuration,
  raw_data,
  callback,
  cl,
  limit = NULL
)
}
\description{
Generate phenotypes based on 3 character ICD10 codes.
}
