% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{generator_self_reported_diseases}
\alias{generator_self_reported_diseases}
\title{Generate self reported diseases.}
\usage{
generator_self_reported_diseases(
  configuration,
  raw_data,
  callback,
  cl,
  limit = NULL
)
}
\description{
Generate self reported diseases.
}
