% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_biomarker_connection}
\alias{get_biomarker_connection}
\title{Get a connection to the UK Biobank biomarkers temporary DB.}
\usage{
get_biomarker_connection(password)
}
\value{
A database connection object
}
\description{
Get a connection to the UK Biobank biomarkers temporary DB.
}
\examples{
\dontrun{
get_biomarker_connection("my_password")
}

}
