% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_covariables_hcn4_project}
\alias{get_covariables_hcn4_project}
\title{Get the age and sex from the same dataset as the HCN4 paper.}
\usage{
get_covariables_hcn4_project(con)
}
\arguments{
\item{con}{A connection to the UK Biobank}
}
\value{
A dataframe of covariables.
}
\description{
Get the age and sex from the same dataset as the HCN4 paper.
}
