% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_full_records}
\alias{get_full_records}
\title{Get all hospitalization and/or death records from the UKB databse.}
\usage{
get_full_records(con, include_secondary_hospit, include_death_records)
}
\description{
Get all hospitalization and/or death records from the UKB databse.
}
