% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{get_icd10_group}
\alias{get_icd10_group}
\title{ICD10 code to group.}
\usage{
get_icd10_group(icd10)
}
\arguments{
\item{icd10}{The code.}
}
\value{
A group name and a unique string representing the range as a list
        e.g. list(range="I20-I25", "Ischemic heart diseases")
}
\description{
The groups are sets of contiguous ICD10 codes. For example, I20-I25 
consists of ischemic heart diseases.
They need to be hard-coded somewhere so that codes can be grouped
automatically during a pheWAS.
}
