% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_icd10_with_at_least_n_cases}
\alias{get_icd10_with_at_least_n_cases}
\title{Get a vec of icd10 codes with at least n occurences in hospitalization data.}
\usage{
get_icd10_with_at_least_n_cases(con, n, include_secondary_hospit = T,
  include_death_records = T)
}
\arguments{
\item{n}{The minimum number of cases for inclusion.}
}
\value{
A list of ICD10 codes.
}
\description{
This relies on the "full hesin" table.
}
