% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_ukb_connection}
\alias{get_ukb_connection}
\title{Get a connection to the UK Biobank database.}
\usage{
get_ukb_connection(password)
}
\value{
A database connection object
}
\description{
Get a connection to the UK Biobank database.
}
\examples{
ukb_connect("my_password")

}
