% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{infer_icd10_level}
\alias{infer_icd10_level}
\title{Infer the level of an ICD10 code.}
\usage{
infer_icd10_level(icd10)
}
\arguments{
\item{icd10}{The code.}
}
\value{
An integer with the level of the code as described above.
}
\description{
Level 0 is the chapter (e.g. I)
Level 1 is for the next two numbers. (e.g. I20)
Level 2 is for codes that are more precise. (e.g. I23.8)
}
