% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas.R, R/stats.R
\name{internal_do_linear}
\alias{internal_do_linear}
\title{Do the linear regression.}
\usage{
internal_do_linear(y, configuration)

internal_do_linear(y, configuration)
}
\arguments{
\item{y}{A data.frame with two columns: sample_id and value (assumed to be
provided in that order).}

\item{configuration}{A configuration object containing information on the
analysis as well as the covariables matrix.}

\item{y}{A data.frame with two columns: sample_id and value (assumed to be
provided in that order).}

\item{configuration}{A configuration object containing information on the
analysis as well as the covariables matrix.}
}
\description{
Do the linear regression for the current biomarker.

Do the linear regression for the current biomarker.
}
