% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas.R, R/stats.R
\name{internal_do_logistic}
\alias{internal_do_logistic}
\title{Do the logistic regression.}
\usage{
internal_do_logistic(cur_cases, configuration)

internal_do_logistic(cur_cases, configuration)
}
\arguments{
\item{cur_cases}{A column data frame of sample IDs to consider to be cases.}

\item{configuration}{A configuration object containing information on the
analysis as well as the covariables matrix.}

\item{cur_cases}{A column data frame of sample IDs to consider to be cases.}

\item{configuration}{A configuration object containing information on the
analysis as well as the covariables matrix.}
}
\description{
Do the actual regression from the current cases and the prepared covariates
dataframe.

Do the actual regression from the current cases and the prepared covariates
dataframe.
}
\seealso{
fastglm

fastglm
}
