% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{internal_get_from_stmt}
\alias{internal_get_from_stmt}
\title{Create a SQL string with an appropriate FROM statement to obtain cases.}
\usage{
internal_get_from_stmt(include_secondary_hospit, include_death_records)
}
\arguments{
\item{include_secondary_hospit}{Include the secondary hospitalization codes.}

\item{include_death_records}{Include the causes of death.}
}
\value{
A string containing the FROM statement
}
\description{
Specifically the generated table will have two columns: eid and diag_icd10
from the appropriate HESIN table (primary or primary+secondary) and with
or without the death records.
}
