% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas.R
\name{internal_run_biomarker_phewas}
\alias{internal_run_biomarker_phewas}
\title{Run a pheWAS on biomarker traits.}
\usage{
internal_run_biomarker_phewas(configuration, con)
}
\arguments{
\item{configuration}{A pheWAS configuration object.}

\item{con}{A connection to the biomarker DB.}
}
\description{
Run a pheWAS on biomarker traits.
}
