% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas.R
\name{run_3_char_pheWAS}
\alias{run_3_char_pheWAS}
\title{Run a 3 character ICD10 code pheWAS.}
\usage{
run_3_char_pheWAS(configuration, con, all_cases, codes)
}
\arguments{
\item{configuration}{A pheWAS configuration object.}

\item{con}{A connection to the DB.}

\item{all_cases}{A dataframe with all of the EMR or death record data.}

\item{codes}{All ICD10 codes that were found in the DB with enough cases.}
}
\description{
Run a 3 character ICD10 code pheWAS.
}
