% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas.R
\name{run_block_phewas}
\alias{run_block_phewas}
\title{Helper function to run a pheWAS on blocks of ICD10 codes.}
\usage{
run_block_phewas(configuration, con, all_cases, codes)
}
\arguments{
\item{configuration}{A pheWAS configuration object.}

\item{con}{A connection to the DB.}

\item{all_cases}{A dataframe with all of the EMR or death record data.}

\item{codes}{All ICD10 codes that were found in the DB with enough cases.}
}
\description{
Helper function to run a pheWAS on blocks of ICD10 codes.
}
