% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas.R
\name{run_naive_phewas}
\alias{run_naive_phewas}
\title{Run a naive pheWAS}
\usage{
run_naive_phewas(configuration, con, all_cases)
}
\arguments{
\item{configuration}{A pheWAS configuration object.}

\item{con}{A connection to the DB.}

\item{all_cases}{A dataframe with all of the EMR or death record data.}
}
\description{
Run a naive pheWAS
}
