% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas.R
\name{runBiomarkerPheWAS}
\alias{runBiomarkerPheWAS}
\title{Main function to run a pheWAS on biomarker data.}
\usage{
runBiomarkerPheWAS(con, configuration)
}
\arguments{
\item{con}{A database connection to the Biomarker UK Biobank database.}

\item{configuration}{A configuration object.}
}
\description{
Main function to run a pheWAS on biomarker data.
}
