% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{strip_icd}
\alias{strip_icd}
\title{Strip dots from the ICD10 code (e.g. I20.0 -> I20).}
\usage{
strip_icd(icd10)
}
\arguments{
\item{icd10}{The code.}
}
\value{
An integer with the level of the code as described above.
}
\description{
Strip dots from the ICD10 code (e.g. I20.0 -> I20).
}
