% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas.R
\name{tidy_fastglm}
\alias{tidy_fastglm}
\title{A "copy" of broom::tidy for fastglm objects.}
\usage{
tidy_fastglm(fit)
}
\description{
The goal is to achieve results close enough or identical to 
format_logistic_fit so that the use of fastglm is mostly transparent.
}
