% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R, R/phewas.R
\name{variable_of_interest_index}
\alias{variable_of_interest_index}
\title{Predicate generator function for variable of interest filtering based on
parameter indices.}
\usage{
variable_of_interest_index(idx = 2)

variable_of_interest_index(idx = 2)
}
\description{
Predicate generator function for variable of interest filtering based on
parameter indices.

Predicate generator function for variable of interest filtering based on
parameter indices.
}
