% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R, R/phewas.R
\name{variable_of_interest_name}
\alias{variable_of_interest_name}
\title{Predicate generator function for variable of interest filtering based on
parameter names.}
\usage{
variable_of_interest_name(name)

variable_of_interest_name(name)
}
\description{
Predicate generator function for variable of interest filtering based on
parameter names.

Predicate generator function for variable of interest filtering based on
parameter names.
}
