% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas.R
\docType{class}
\name{Configuration-class}
\alias{Configuration-class}
\alias{Configuration}
\title{A configuration class to parametrize PheWAS analyses.}
\description{
A configuration class to parametrize PheWAS analyses.
}
