% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas.R
\name{create_configuration}
\alias{create_configuration}
\title{Construction function to create configuration objects for pheWAS analyses.}
\usage{
create_configuration(include_secondary_hospit = TRUE,
  include_death_records = TRUE, min_num_cases = 50, ncpus = 10,
  voi_idx = NULL, voi_name = NULL, voi_predicate = NULL, xs = NULL,
  model_rhs = "", output_prefix = "phewas", use_fastglm = TRUE,
  do_continuous = TRUE, do_binary = TRUE)
}
\description{
This does mostly parameter checking.
}
