% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{do_logistic}
\alias{do_logistic}
\title{Do the logistic regression.}
\usage{
do_logistic(configuration, data)
}
\description{
Do the actual regression from the current cases and the prepared covariates
dataframe.
}
\seealso{
fastglm
}
