% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas.R
\name{internal_do_logistic}
\alias{internal_do_logistic}
\title{Do the logistic regression.}
\usage{
internal_do_logistic(cur_cases, xs_df)
}
\arguments{
\item{cur_cases}{A column data frame of sample IDs to consider to be cases.}

\item{xs_df}{A matrix of covariates to be included in the multivariate
regression.}
}
\description{
Do the actual regression from the current cases and the prepared covariates
dataframe.
}
