% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas.R
\name{runPheWAS}
\alias{runPheWAS}
\title{Main function to run a pheWAS.}
\usage{
runPheWAS(configuration)
}
\arguments{
\item{configuration}{A configuration object.}

\item{con}{A database connection to the UK Biobank (StatGen format).}
}
\description{
This function takes a configuration object and dispatches to the relevant
function.
}
